refactoring

+++ backend/common/src/main/java/com/common/data/activity/LogActivity.java
@@ -17,6 +17,6 @@ import org.springframework.data.mongodb.core.mapping.Document;
         private String email;
         private String log;
         private Date  date;
-        private Long usePoints;    
+        private Integer usePoints;
     
 }
